+++ packages/signers/signers-common/src/v4.ts
@@ -77,7 +77,7 @@ export const createV4Tx = (
   signingType?: "Ecdsa" | "Ed25519" | "Sr25519",
 ) => {
   const [signerType, addressPrefix] = getSignerType(metadata)
-  const preResult = mergeUint8(
+  const preResult = mergeUint8([
     versionCodec({ signed: true, version: 4 }),
     // converting it to a `MultiAddress` enum, where the index 0 is `Id(AccountId)`
     signerType === SignerType.Ethereum
@@ -88,6 +88,6 @@ export const createV4Tx = (
       : new Uint8Array([signingTypeId[signingType], ...signed]),
     ...extra,
     callData,
-  )
-  return mergeUint8(compact.enc(preResult.length), preResult)
+  ])
+  return mergeUint8([compact.enc(preResult.length), preResult])
 }