+++ packages/utils/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Add `jsonPrint` function
 
+### Changed
+
+- `mergeUint8` now takes an `Array<Uint8Array>`. Soft deprecate old overload.
+
 ## 0.1.2 - 2024-10-05
 
 ### Fixed