+++ packages/utils/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.2.0 - 2025-05-30
+
 ### Added
 
 - Add `jsonPrint` function